requirements.txt
pyrogram==2.0.106
tgcrypto>=1.2.5
fastapi==0.115.4
uvicorn[standard]==0.32.0
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.2
redis==5.0.8
